rename util directory to utils
